requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp>=3.8.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
psutil==5.9.6
Flask==3.0.0
waitress==3.0.0
gunicorn==21.2.0